Updated parameter parsing in the command class.
Added the functionality for the command class to support optional
parameters in the form of <prefix><name>=<value> where it is a single
string rather two seperate items. Fixed one of the unit tests which was
failing because it was incorrect it should have been testing that an
exception was being thrown but it wasn't.